+++ lib/bricks/dsl.rb
@@ -8,12 +8,20 @@ module Bricks
       build(klass).generate
     end
 
+    def build?(klass)
+      build(klass).generate(:search => true)
+    end
+
     def create(klass)
       builder(klass, true)
     end
 
     def create!(klass)
       create(klass).generate
+    end
+
+    def create?(klass)
+      create(klass).generate(:search => true)
     end
 
     def builder(klass, save)