+++ History.md
@@ -5,3 +5,4 @@ History
 -----
 
 * Blocks passed to attributes now optionally take a second argument (the builder parent).
+* You can now use `?` like `!`, but it will search for an existing record first.